Use one less than the motif length as the sequence radius to inspect (slams #3 shut)
If a motif begins right after a variant and we use the motif length
as the radius, it will be detected when we read the flanking sequence
despite having nothing to do with the variant. This kind of situation
is noise. By using motif_length - 1, we get the longest case where the
first element of a motif is affected by a sequence and no noise.